requirements.txt
discord.py==1.5.1
requests==2.25.1
dblpy==0.4.0
asyncpg==0.21.0
